+++ swype-client/src/components/Modal/InfoModal.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
-import Button from './Button';
 import Quokka from '../../assets/Quokka.png';
 import { useState } from 'react';
+import Button from '../common/Button';
 
 const Wrapper = styled.div`
   display: flex;
