+++ swype-client/src/components/Modal/QuizModal.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Leeca from '../../assets/LeeCa.png';
-import { Container } from '../share/utils/GlobalStyle';
+import { Container } from '../../share/utils/GlobalStyle';
 
 export default function QuizModal() {
   return (
